requirements.txt
django
djangorestframework
celery
django-celery-beat